Finish 07-flex-layout-2 and foundations-flex
Finished this last exercise, really fun. Flexbox is awesome.
Took me a while to get the cards correct, I was for some reason
trying to use align-items center on the cards wrapper I made, which
was causing unintended behavior when wrapping, but I deleted it
and everything was perfect. Happy to be done with flex exercises and
excited for the next project and JS after.